+++ src/main/java/web/Service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserService {
     List<User> allUsers();
     void add(User user);
-    void delete(Integer id);
+    void delete(long id);
     void edit(User user);
-    User getById(Integer id);
+    User getById(long id);
 }
